+++ lib/statewide_parser.rb
@@ -1,71 +1,7 @@
 require_relative 'helper'
 
-module CSVParser
+module StatewideParser
   include Helper
-
-  def district_repo_parser(file_tree)
-    all_data = []
-    files = file_tree.values.reduce({}, :merge!)
-    i = files.count
-    path_counter = 0
-    i.times do
-      filepath = files[files.keys[path_counter]]
-       CSV.foreach(filepath, headers: true, header_converters: :symbol).each do |row|
-         district = { :name => row[:location].upcase}
-        if row[:location] && all_data.include?(district) == false
-          all_data << district
-        end
-      end
-      path_counter += 1
-    end
-    all_data
-  end
-
-  def parsed_enrollment_data(filepath)
-    CSV.foreach(filepath, headers: true, header_converters: :symbol).map do |row|
-      { name: row[:location].upcase, row[:timeframe].to_i => row[:data].to_f}
-    end
-  end
-
-  def group_names(all_parsed_data)
-    all_parsed_data.group_by do |row|
-      row[:name].upcase
-    end
-  end
-
-  def year_merge(grade_level, enrollment_by_year)
-    enrollment_by_year.map do |name, years|
-      year_data = years.reduce({}, :merge)
-      year_data.delete(:name)
-      {:name => name, grade_level => year_data }
-    end
-  end
-
-  def combined_enrollment_by_year(all_data)
-    group_names(all_data).values.map do |district_set|
-      district_set.reduce({}, :merge)
-    end
-  end
-
-  def enrollment_repo_parser(file_tree)
-    path_counter = 0
-    all_data = []
-    i = file_tree.values[0].values.count
-    i.times do
-      filepath = file_tree.values[0].values[path_counter]
-      enrollment_by_year = group_names(parsed_enrollment_data(filepath))
-      if file_tree.values[0].keys[path_counter] == :kindergarten
-        grade_level = :kindergarten_participation
-      else
-        grade_level = file_tree.values[0].keys[path_counter]
-      end
-      year_merge(grade_level, enrollment_by_year).each do |data_set|
-        all_data << (data_set)
-      end
-      path_counter += 1
-    end
-    combined_enrollment_by_year(all_data)
-  end
 
   def parsed_test_data(file_tree, path_counter)
     filepath = file_tree.values[0].values[path_counter]
