+++ lib/statewide_test.rb
@@ -4,13 +4,16 @@ require_relative 'custom_errors'
 
 class StatewideTest
   include Helper
-  attr_reader :attributes,
-              :parameter
+  attr_reader       :attributes,
+                    :parameter
 
     def initialize(attributes)
       @attributes = attributes
       @parameter = {3 => :third_grade, 8 => :eighth_grade,
-                    :asian => :asian, :pacific_islander => :pacific_islander, :black => :black, :hispanic => :hispanic, :native_american => :native_american, :two_or_more => :two_or_more, :white => :white, :reading => :reading, :writing => :writing, :math => :math}
+                    :asian => :asian, :pacific_islander => :pacific_islander,
+                    :black => :black, :hispanic => :hispanic, :native_american => :native_american,
+                    :two_or_more => :two_or_more, :white => :white, :reading => :reading,
+                    :writing => :writing, :math => :math}
     end
 
     def name
@@ -44,4 +47,5 @@ class StatewideTest
       raise UnknownDataError unless valid_data(subject, race, year)
       attributes[race][year][subject]
     end
+
 end
